refactor: make BlogPosts a common component
